feat(study-room-scraper): ✨ define prisma schema
